% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ebrd.R
\name{list_ebrd_groups}
\alias{list_ebrd_groups}
\title{EBRD regional groupings}
\usage{
list_ebrd_groups()
}
\value{
A character vector of grouping names.
}
\description{
Returns the names of the traditional EBRD regional groupings (\code{coo_group})
that can be passed to the \code{group} argument of \code{\link[=list_ebrd]{list_ebrd()}}.
}
\examples{
list_ebrd_groups()
}
\seealso{
\code{\link[=list_ebrd]{list_ebrd()}}
}
