% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebrdify.R
\name{ebrdify}
\alias{ebrdify}
\title{EBRD Country Classification}
\usage{
ebrdify(data = NULL, var, var_format = NULL)
}
\arguments{
\item{data}{A data frame containing the variable to classify, or NULL if using a vector input.}

\item{var}{A string specifying the name of the variable in \code{data} that contains the country codes, or a vector of country codes.}

\item{var_format}{A string specifying the format of the country codes in \code{var}. It can be "country.name", "iso3c", or "iso2c". If NULL, the function will attempt to detect the format.}
}
\value{
A data frame with two new variables: \code{ebrd} and \code{coo_group}, and prints out any unmatched entries.
}
\description{
This function classifies countries based on their EBRD status and region.
}
\examples{
# Using a data frame
data <- data.frame(country_code = c("KAZ", "CZE", "GRC", "ARM", "ALB", "EGY", "USA", "CAN"))
ebrdified_data <- ebrdify(data, "country_code", var_format = "iso3c")
print(ebrdified_data)

# Using a vector
country_vector <- c("KAZ", "CZE", "GRC", "ARM", "ALB", "EGY", "USA", "CAN")
ebrdified_vector <- ebrdify(var = country_vector, var_format = "iso3c")
print(ebrdified_vector)

# Using a data frame with fake country names
data_fake_names <- data.frame(country_name = c("Kazakhstan",
                                               "Czechia",
                                               "Narnia",
                                               "Armenia",
                                               "Albania",
                                               "Wakanda",
                                               "United States",
                                               "Canada"))
ebrdified_data_fake_names <- ebrdify(data_fake_names, "country_name")
print(ebrdified_data_fake_names)
}
