% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebrdify.R
\name{ebrdify}
\alias{ebrdify}
\title{EBRD Country Classification}
\usage{
ebrdify(data = NULL, var, var_format = NULL)
}
\arguments{
\item{data}{A data frame containing the variable to classify, or NULL if using a vector input.}

\item{var}{A string specifying the name of the variable in \code{data} that contains the country codes, or a vector of country codes.}

\item{var_format}{A string specifying the format of the country codes in \code{var}. It can be "country.name", "iso3c", or "iso2c". If NULL, the function will attempt to detect the format.}
}
\value{
A data frame with four new variables: \code{ebrd}, \code{coo_group}, \code{eu_ebrd}, and \code{coo_group_alt}, and prints out any unmatched entries.
\itemize{
\item \code{ebrd}: A binary variable indicating whether the country is an EBRD country of operation (1 = EBRD COO, 0 = Non-COO).
\item \code{coo_group}: A variable classifying the country into specific EBRD country groupings.
\item \code{eu_ebrd}: A binary variable indicating whether the country is both an EBRD country of operation and an EU member (1 = EU & EBRD, 0 = otherwise).
\item \code{coo_group_alt}: An alternative classification of countries into broader categories
\item \code{ebrd_shareholder}: A binary variable indicating whether the country is an EBRD shareholder (1 = Shareholder, 0 = Non-Shareholder).
}
}
\description{
This function classifies countries based on their EBRD status, region, and EU membership.
}
\examples{
# Using a data frame
data <- data.frame(country_code = c("KAZ", "CZE", "GRC", "ARM", "ALB", "EGY", "USA", "CAN"))
ebrdified_data <- ebrdify(data, "country_code", var_format = "iso3c")
print(ebrdified_data)

# Using a vector
country_vector <- c("KAZ", "CZE", "GRC", "ARM", "ALB", "EGY", "USA", "CAN")
ebrdified_vector <- ebrdify(var = country_vector, var_format = "iso3c")
print(ebrdified_vector)

# Using a data frame with fake country names
data_fake_names <- data.frame(country_name = c("Kazakhstan",
                                               "Czechia",
                                               "Narnia",
                                               "Armenia",
                                               "Albania",
                                               "Wakanda",
                                               "Kosovo",
                                               "United States",
                                               "Canada"))
ebrdified_data_fake_names <- ebrdify(data_fake_names, "country_name")
print(ebrdified_data_fake_names)

}
